:truck: rename: MemberProfileService.java 를 application 폴더로 이동 #26
